requirements.txt
absl-py==0.12.0
appdirs==1.4.4
asgiref==3.3.4
astunparse==1.6.3
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
blinker==1.4
boto3==1.17.102
botocore==1.20.102
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==5.0.9
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.4
django-extensions==3.1.3
djangorestframework==3.12.4
dynamodb-json==1.3
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.10.1
jmespath==0.10.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
Markdown==3.3.4
MarkupSafe==1.1.0
marshmallow==3.12.2
marshmallow-pynamodb==0.8.0
matplotlib-inline==0.1.2
more-itertools==4.2.0
mysqlclient==2.0.3
netifaces==0.10.4
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
protobuf==3.17.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.9.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynamodb==5.1.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
s3transfer==0.4.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
sos==4.1
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
traitlets==5.0.5
Twisted==18.9.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.4.7
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==1.0.0
zope.interface==4.7.1
